step 10

+++ app/src/main/java/hexlet/code/service/TaskStatusService.java
@@ -7,6 +7,9 @@ import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.TaskStatusMapper;
 import hexlet.code.model.TaskStatus;
 import hexlet.code.repository.TaskStatusRepository;
+import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,41 +17,52 @@ import java.util.List;
 
 @Service
 public class TaskStatusService {
+    private static final Logger LOGGER = LoggerFactory.getLogger(TaskService.class);
+    private final TaskStatusRepository taskStatusRepository;
+    private final TaskStatusMapper taskStatusMapper;
 
     @Autowired
-    private TaskStatusRepository taskStatusRepository;
-    @Autowired
-    private TaskStatusMapper taskStatusMapper;
-
-    public TaskStatusDTO getTaskStatusById(Long id) {
-        TaskStatus taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Task status with id '%d' not found".formatted(id)));
-        return taskStatusMapper.map(taskStatus);
+    public TaskStatusService(TaskStatusRepository taskStatusRepository, TaskStatusMapper taskStatusMapper) {
+        this.taskStatusRepository = taskStatusRepository;
+        this.taskStatusMapper = taskStatusMapper;
     }
 
-    public List<TaskStatusDTO> getAll() {
-        List<TaskStatus> taskStatusList = taskStatusRepository.findAll();
-        return taskStatusList.stream()
+    public List<TaskStatusDTO> getAllTaskStatuses() {
+        return taskStatusRepository.findAll().stream()
                 .map(taskStatusMapper::map)
                 .toList();
     }
 
-    public TaskStatusDTO create(TaskStatusCreateDTO taskStatusCreateDTO) {
+    public TaskStatusDTO getTaskStatusById(Long id) {
+        TaskStatus taskStatus = taskStatusRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Task status with id '%d' not found", id));
+        return taskStatusMapper.map(taskStatus);
+    }
+
+    @Transactional
+    public TaskStatusDTO createTaskStatus(TaskStatusCreateDTO taskStatusCreateDTO) {
         TaskStatus taskStatus = taskStatusMapper.map(taskStatusCreateDTO);
         taskStatus = taskStatusRepository.save(taskStatus);
+        LOGGER.info("Created task status with id: {}", taskStatus.getId());
         return taskStatusMapper.map(taskStatus);
     }
 
-    public TaskStatusDTO update(Long id, TaskStatusUpdateDTO taskStatusUpdateDTO) {
-        TaskStatus taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Task status with id '%d' not found".formatted(id)));
+    @Transactional
+    public TaskStatusDTO updateTaskStatus(Long id, TaskStatusUpdateDTO taskStatusUpdateDTO) {
+        TaskStatus taskStatus = taskStatusRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Task status with id '%d' not found", id));
         taskStatusMapper.update(taskStatusUpdateDTO, taskStatus);
+        LOGGER.info("Updated task status with id: {}", id);
         taskStatusRepository.save(taskStatus);
-
         return taskStatusMapper.map(taskStatus);
     }
 
+    @Transactional
     public void deleteById(Long id) {
+        if (!taskStatusRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Task status with id '%d' not found", id);
+        }
         taskStatusRepository.deleteById(id);
+        LOGGER.info("Deleted task status with id: {}", id);
     }
 }
